Refactored the handlers into controllers
