+++ helpers/controllerHelper.go
@@ -1,0 +1,21 @@
+//Package helpers contains helper methods that support other packages.
+package helpers
+
+import (
+	"net/http"
+	"regexp"
+)
+
+var validPath = regexp.MustCompile("^/(edit|save|view)/([a-zA-Z0-9]+)$")
+
+//MakeController tests for valid Request path before returning a supported handler.
+func MakeController(fn func(http.ResponseWriter, *http.Request, string)) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		m := validPath.FindStringSubmatch(r.URL.Path)
+		if m == nil {
+			http.NotFound(w, r)
+			return
+		}
+		fn(w, r, m[2])
+	}
+}
