+++ higo.go
@@ -2,69 +2,19 @@
 package main
 
 import (
-	"html/template"
 	"net/http"
-	"regexp"
 
-	"github.com/mikeyb214/higo/models"
+	"github.com/mikeyb214/higo/controllers"
+	"github.com/mikeyb214/higo/helpers"
 )
-
-var validPath = regexp.MustCompile("^/(edit|save|view)/([a-zA-Z0-9]+)$")
-var templates = template.Must(template.ParseFiles("./views/edit.gohtml", "./views/view.gohtml"))
-
-func viewHandler(w http.ResponseWriter, r *http.Request, title string) {
-	p, err := models.LoadPage(title)
-	if err != nil {
-		http.Redirect(w, r, "/edit/"+title, http.StatusFound)
-		return
-	}
-	renderTemplate(w, "view", p)
-}
-
-func editHandler(w http.ResponseWriter, r *http.Request, title string) {
-	p, err := models.LoadPage(title)
-	if err != nil {
-		p = &models.Page{Title: title}
-	}
-	renderTemplate(w, "edit", p)
-}
-
-func saveHandler(w http.ResponseWriter, r *http.Request, title string) {
-	body := r.FormValue("body")
-	p := &models.Page{Title: title, Body: []byte(body)}
-	err := p.SavePage()
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-		return
-	}
-	http.Redirect(w, r, "/view/"+title, http.StatusFound)
-}
-
-func makeHandler(fn func(http.ResponseWriter, *http.Request, string)) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		m := validPath.FindStringSubmatch(r.URL.Path)
-		if m == nil {
-			http.NotFound(w, r)
-			return
-		}
-		fn(w, r, m[2])
-	}
-}
-
-func renderTemplate(w http.ResponseWriter, tmpl string, p *models.Page) {
-	err := templates.ExecuteTemplate(w, tmpl+".gohtml", p)
-	if err != nil {
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-	}
-}
 
 func main() {
 	http.HandleFunc("/", func(w http.ResponseWriter, r *http.Request) {
 		http.ServeFile(w, r, "./public/index.html")
 	})
-	http.HandleFunc("/view/", makeHandler(viewHandler))
-	http.HandleFunc("/edit/", makeHandler(editHandler))
-	http.HandleFunc("/save/", makeHandler(saveHandler))
+	http.HandleFunc("/view/", helpers.MakeController(controllers.ViewController))
+	http.HandleFunc("/edit/", helpers.MakeController(controllers.EditController))
+	http.HandleFunc("/save/", helpers.MakeController(controllers.SaveController))
 
 	http.ListenAndServe(":8080", nil)
 }
